Refactor step structure [WIP]
